+++ shortlink/src/index.ts
@@ -1,8 +1,7 @@
-import installScript from "./install.sh";
 import { handleRequest } from "./handler";
 
 export default {
   fetch(request: Request): Response {
-    return handleRequest(request, installScript);
+    return handleRequest(request);
   }
 } satisfies ExportedHandler<Env>;